+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - Click [here](https://r35007.github.io/Mock-Server/) for mock server documentation.
 - Click [here](https://github.com/R35007/Mock-Server/blob/main/CHANGELOG.md) to see more changes reg the `@r35007/mock-server` package
 
+## v12.1.0
+
+- Updated `@r35007/mock-server@12.0.0`.
+- samples update.
+- Added - Right click on mock server config `.js` or `.json` file and select `Set as Mock Server Config` to set the mock server config in vscode config
+
 ## v12.0.0
 
 - Updated `@r35007/mock-server@12.0.0`.